+++ src/main/scala/millfork/assembly/mos/opt/ZeropageRegisterOptimizations.scala
@@ -13,6 +13,8 @@ object ZeropageRegisterOptimizations {
 
   val functionsThatUsePseudoregisterAsInput: Map[String, Set[Int]] = Map(
     "__mul_u8u8u8" -> Set(0, 1),
+    "__mod_u8u8u8u8" -> Set(0, 1),
+    "__div_u8u8u8u8" -> Set(0, 1),
     "__mul_u16u8u16" -> Set(0, 1, 2),
     "__adc_decimal" -> Set(2, 3),
     "__sbc_decimal" -> Set(2, 3),