+++ src/main/scala/millfork/assembly/z80/opt/CoarseFlowAnalyzer.scala
@@ -25,7 +25,7 @@ object CoarseFlowAnalyzer {
     val preservesB: Set[String] = Set("__mul_u8u8u8")
     val preservesC: Set[String] = if (z80) Set("__mul_u8u8u8") else Set()
     val preservesD: Set[String] = Set()
-    val preservesE: Set[String] = Set()
+    val preservesE: Set[String] = Set("__divmod_u16u8u16u8")
     val preservesH: Set[String] = Set("__mul_u8u8u8")
     val preservesL: Set[String] = Set("__mul_u8u8u8")