+++ src/test/scala/millfork/test/ByteMathSuite.scala
@@ -314,8 +314,8 @@ class ByteMathSuite extends FunSuite with Matchers with AppendedClues {
          | void main () {
          |  byte a
          |  a = f()
-         |  //output_q1 = a / $y
-         |  //output_m1 = a %% $y
+         |  output_q1 = a / $y
+         |  output_m1 = a %% $y
          |  output_q2 = a
          |  output_m2 = a
          |  output_q2 /= $y
@@ -324,8 +324,53 @@ class ByteMathSuite extends FunSuite with Matchers with AppendedClues {
          | byte f() {return $x}
           """.
         stripMargin) { m =>
-//      m.readByte(0xc000) should equal(x / y) withClue s"$x / $y"
-//      m.readByte(0xc001) should equal(x % y) withClue s"$x %% $y"
+      m.readByte(0xc000) should equal(x / y) withClue s"$x / $y"
+      m.readByte(0xc001) should equal(x % y) withClue s"$x %% $y"
+      m.readByte(0xc002) should equal(x / y) withClue s"$x / $y"
+      m.readByte(0xc003) should equal(x % y) withClue s"$x %% $y"
+    }
+  }
+
+  test("Byte division 2") {
+    divisionCase2(0, 1)
+    divisionCase2(1, 1)
+    divisionCase2(2, 1)
+    divisionCase2(250, 1)
+    divisionCase2(0, 3)
+    divisionCase2(0, 5)
+    divisionCase2(1, 5)
+    divisionCase2(6, 5)
+    divisionCase2(73, 5)
+    divisionCase2(75, 5)
+    divisionCase2(42, 11)
+  }
+
+  private def divisionCase2(x: Int, y: Int): Unit = {
+    EmuCrossPlatformBenchmarkRun(Cpu.Mos, Cpu.Z80, Cpu.Intel8080, Cpu.Sharp, Cpu.Intel8086)(
+      s"""
+         | import zp_reg
+         | byte output_q1 @$$c000
+         | byte output_m1 @$$c001
+         | byte output_q2 @$$c002
+         | byte output_m2 @$$c003
+         | void main () {
+         |  byte a
+         |  byte b
+         |  a = f()
+         |  b = g()
+         |  output_q1 = a / b
+         |  output_m1 = a %% b
+         |  output_q2 = a
+         |  output_m2 = a
+         |  output_q2 /= b
+         |  output_m2 %%= b
+         | }
+         | byte f() = $x
+         | noinline byte g() = $y
+          """.
+        stripMargin) { m =>
+      m.readByte(0xc000) should equal(x / y) withClue s"$x / $y"
+      m.readByte(0xc001) should equal(x % y) withClue s"$x %% $y"
       m.readByte(0xc002) should equal(x / y) withClue s"$x / $y"
       m.readByte(0xc003) should equal(x % y) withClue s"$x %% $y"
     }